refactor: add hide functionality to the registration page (#41)
