+++ src/pages/LoginPage/LoginPage.module.css
@@ -18,33 +18,6 @@
   text-align: center;
 }
 
-.password_box {
-  color: var(--dark-primary);
-  display: flex;
-  flex-direction: column;
-  gap: 5px;
-}
-
-.label_box {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.password_hide {
-  all: unset;
-  cursor: pointer;
-  transition: 0.25s;
-
-  &:hover {
-    color: var(--primary);
-  }
-}
-
-.password_error {
-  white-space: pre-wrap;
-}
-
 .buttons {
   display: flex;
   gap: 10px;
@@ -63,4 +36,5 @@
 
 .error {
   color: var(--error);
+  white-space: pre-wrap;
 }
