+++ src/pages/RegistrationPage/RegistrationPage.module.css
@@ -36,4 +36,5 @@
 
 .error {
   color: var(--error);
+  white-space: pre-wrap;
 }
